add jwt middleware
