--- conf/config.js
@@ -1,22 +1,0 @@
-const env = process.env
-
-module.exports = {
-  serverPort: env.serverPort || 3000,
-
-  mongoHost: env.mongoHost || '127.0.0.1',
-  mongoDatabase: env.mongoDatabase || 'blog',
-  mongoPort: env.mongoPort || 27017,
-
-  redisHost: env.redisHost || '127.0.0.1',
-  redisPort: env.redisPort || 6379,
-  redisPassword: env.redisPassword || '',
-
-  tokenSecret: env.tokenSecret || 'test',
-  tokenExpiresIn: env.tokenExpiresIn || '7d',
-
-  defaultAdminName: env.defaultAdminName || 'admin',
-  defaultAdminPassword: env.defaultAdminPassword || '123456',
-
-  rongyun_key: 'lmxuhwagxgt9d',
-  rongyun_secret: 'NpbRLWPxB79',
-}
