--- conf/option.js
@@ -1,59 +1,0 @@
-module.exports = [
-  {
-    'key': 'analyzeCode',
-    'value': ''
-  },
-  {
-    'key': 'commentType',
-    'value': 'disqus'
-  },
-  {
-    'key': 'commentName',
-    'value': ''
-  },
-  {
-    'key': 'description',
-    'value': ''
-  },
-  {
-    'key': 'faviconUrl',
-    'value': '/static/favicon.ico'
-  },
-  {
-    'key': 'logoUrl',
-    'value': '/static/logo.png'
-  },
-  {
-    'key': 'githubUrl',
-    'value': ''
-  },
-  {
-    'key': 'keywords',
-    'value': '',
-    'desc': '网站关键字'
-  },
-  {
-    'key': 'miitbeian',
-    'value': ''
-  },
-  {
-    'key': 'numPerPage',
-    'value': ''
-  },
-  {
-    'key': 'siteUrl',
-    'value': ''
-  },
-  {
-    'key': 'title',
-    'value': ''
-  },
-  {
-    'key': 'weiboUrl',
-    'value': ''
-  },
-  {
-    'key': 'twoFactorAuth',
-    'value': ''
-  }
-]
